refactor: eps service
