+++ internal/logic/rollup/eps_service.go
@@ -1,4 +1,4 @@
-package indexer
+package rollup
 
 import (
 	"encoding/json"
